split metrics from runtime API
